Remove comment that appeared from #332 PR
We try to focus on just acknowledging which games are enabled to buy and
run on Steam Linux; Even "Broken overlay." is a bit out of scope but we
allow it as it is a Steam feature that is missing for those titles.

These comments appear inline on SteamDB’s linux page, are seen by users
and devs(!) and might be read - even with the current disclaimers - as
being from us all.

Thanks.